+++ tests/layers_views/views/VivView.spec.js
@@ -6,14 +6,14 @@ import { getVivId } from '../../../src/views/utils';
 
 export const defaultArguments = {
   initialViewState: {
-    height: 500,
-    width: 1000,
     target: [5000, 5000, 0],
     zoom: -4,
     id: 'foo'
   },
   x: 100,
-  y: 50
+  y: 50,
+  height: 500,
+  width: 1000
 };
 
 export function generateViewTests(ViewType, args, linkedViewIds = []) {